# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Go-Xops swagger接口文档",
        "title": "Go-Xops",
        "contact": {
            "name": "pilaoban",
            "url": "https://github.com/jkuup",
            "email": "alphkiee@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "host": "127.0.0.1:9000",
    "paths": {
...
        "/api/v1/api/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "创建api",
                "produces": [
                    "application/json"
                ],
                "summary": "Get /api/v1/api/create",
                "parameters": [
                    {
                        "description": "name, method, path, category, creator, desc",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateApiReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.RespInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.RespInfo"
                        }
                    }
                }
            }
        },
        "/api/v1/api/delete": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据ID批量删除api",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete /api/v1/api/delete",
                "parameters": [
                    {
                        "description": "ids",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.IdsReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.RespInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.RespInfo"
                        }
                    }
                }
            }
        },
        "/api/v1/api/list": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "查看所有API",
                "produces": [
                    "application/json"
                ],
                "summary": "Get /api/v1/api/list",
                "parameters": [
                    {
                        "description": "name, method, path, category, creator, tree",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ApiListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.RespInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.RespInfo"
                        }
                    }
                }
            }
        },
        "/api/v1/api/update/:apiId": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "更新api",
                "produces": [
                    "application/json"
                ],
                "summary": "Get /api/v1/api/update/:apiId",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "apiId",
                        "name": "apiId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.RespInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.RespInfo"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "request.ApiListReq": {
            "type": "object",
            "properties": {
                "all": {
                    "description": "不使用分页",
                    "type": "boolean"
                },
                "category": {
                    "type": "string"
                },
                "creator": {
                    "type": "string"
                },
                "current": {
                    "description": "当前页码",
                    "type": "integer"
                },
                "method": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "pageSize": {
                    "description": "每页显示条数",
                    "type": "integer"
                },
                "path": {
                    "type": "string"
                },
                "total": {
                    "description": "数据总条数",
                    "type": "integer"
                },
                "tree": {
                    "type": "boolean"
                }
            }
        },
...
        "request.CreateApiReq": {
            "type": "object",
            "required": [
                "category",
                "method",
                "name",
                "path"
            ],
            "properties": {
                "category": {
                    "type": "string"
                },
                "creator": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "method": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
...
        "request.IdsReq": {
            "type": "object",
            "properties": {
                "ids": {
                    "description": "传多个id",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "response.RespInfo": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "错误代码代码",
                    "type": "integer"
                },
                "data": {
                    "description": "数据内容",
                    "type": "object"
                },
                "message": {
                    "description": "消息提示",
                    "type": "string"
                },
                "status": {
                    "description": "状态",
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}